chore: fix cortex bundle relative path lookup (#758)
