+++ .github/workflows/cortex-build.yml
@@ -1,4 +1,4 @@
-name: CI Cortex CPP
+name: CI Cortex Release
 
 on:
   push:
